services/job_runner/pyproject.toml
[tool.poetry]
authors = ["Sylvain Lesage <sylvain.lesage@huggingface.co>"]
description = "Worker that refreshes datasets and splits"
name = "job_runner"
version = "0.1.0"

[tool.poetry.dependencies]
Pillow = "^8.4.0"
PyICU = "^2.7.4"
aiohttp = "^3.7.4.post0"
apache-beam = "^2.33.0"
appdirs = "^1.4.4"
bs4 = "^0.0.1"
conllu = "^4.4.1"
datasets = { extras = ["audio", "vision"], version = "^2.1.0" }
gdown = "^4.2.0"
kenlm = { url = "https://github.com/kpu/kenlm/archive/master.zip" }
kss = "^2.6.0"
libcache = { path = "../../libs/libcache", develop = true }
libqueue = { path = "../../libs/libqueue", develop = true }
libutils = { path = "../../libs/libutils", develop = true }
lm-dataformat = "^0.0.20"
lxml = "^4.6.3"
nlp = "^0.4.0"
nltk = "^3.6.5"
openpyxl = "^3.0.9"
psutil = "^5.8.0"
py7zr = "^0.17.4"
pydub = "^0.25.1"
python = "3.9.6"
python-dotenv = "^0.20.0"
rarfile = "^4.0"
requests = "^2.27.1"
sklearn = "^0.0"
tensorflow = "2.8.0"
tfrecord = "^1.14.1"
torchaudio = "^0.10.1"
transformers = "^4.11.3"
trec-car-tools = { path = "../../vendors/trec-car-tools/python3" }
typer = "^0.4.0"
types-psutil = "^5.8.13"
types-requests = "^2.25.11"
wget = "^3.2"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^22.1.0"
flake8 = "^3.9.2"
isort = "^5.9.3"
mypy = "0.812"
poetryup = "^0.3.8"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
safety = "^1.10.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["job_runner"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 119
preview = true

[tool.mypy]
exclude = 'vendors'
strict = true

[[tool.mypy.overrides]]
# could be solved in datasets by adding __add__ to the __init__.py file
implicit_reexport = true
module = "datasets"
